feat: centralize logger configuration
