+++ src/utils/logger.ts
@@ -1,0 +1,12 @@
+import winston from 'winston';
+
+const logger = winston.createLogger({
+    level: process.env.LOG_LEVEL || 'info',
+    format: winston.format.json(),
+    defaultMeta: { service: 'user-service' },
+    transports: [
+        new winston.transports.Console({format: winston.format.simple()})
+    ],
+});
+
+export default logger;
